+++ tests/testbed/013-directory-affinity-ranks-the-near-definition/lib/services/near.rb
@@ -1,0 +1,4 @@
+class Near
+  def perform
+  end
+end